将 jbig2dec 的源码还原为原始代码
并改为使用修改工程配置的方法解决编译问题。
目的是后续升级 jbig2dec 时直接使用上游新代码覆盖即可，无需再合并代码。

jbig2dec 原始代码来源为：
https://github.com/ArtifexSoftware/jbig2dec/tree/a5b9b7cd7b9ed349964eb89883c4b22c410b3398

操作步骤：
将 jbig2dec 原始代码中所有的 .h 覆盖到 JBig2/include/ 目录；
将 .c 文件覆盖到 JBig2/src/ 目录；除了 getopt.c 和 getopt1.c 两个多出来的文件未拷贝。
覆盖后这些文件撤销了之前为解决编译问题而修改的代码。

对于 Visual Studio 的编译错误：
Error C4146 unary minus operator applied to unsigned type, result still unsigned
解决方式为，在 Visual Studio 的
jbig2 Property Pages -> Configuration Properties -> C/C++ -> Command Line -> Additional Options
增加 -w24146 来解决。
Release 和 Debug 都增加。

+++ JBig2/src/jbig2_halftone.c
@@ -83,7 +83,7 @@ jbig2_hd_new(Jbig2Ctx *ctx, const Jbig2PatternDictParams *params, Jbig2Image *im
             /* compose with the REPLACE operator; the source
                will be clipped to the destination, selecting the
                proper sub image */
-            code = jbig2_image_compose(ctx, new->patterns[i], image, -(int32_t)i * (int32_t) HPW, 0, JBIG2_COMPOSE_REPLACE);
+            code = jbig2_image_compose(ctx, new->patterns[i], image, -i * (int32_t) HPW, 0, JBIG2_COMPOSE_REPLACE);
             if (code < 0) {
                 jbig2_error(ctx, JBIG2_SEVERITY_WARNING, JBIG2_UNKNOWN_SEGMENT_NUMBER, "failed to compose image into collective bitmap dictionary");
                 /* new->patterns[i] above succeeded, so release all patterns 0..i */
